social_media_api/requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==6.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
sqlparse==0.5.4
tzdata==2025.2
whitenoise==6.11.0
